fixed getVisisted and moved structs
